requirements.txt
Mezzanine==4.0.1
appnope==0.1.0
beautifulsoup4==4.4.1
bleach==1.4.2
chardet==2.3.0
decorator==4.0.6
Django==1.8.7
django-admin-bootstrapped==2.5.7
django-authtools==1.4.0
django-braces==1.8.1
django-contrib-comments==1.6.2
django-cors-headers==1.1.0
django-crispy-forms==1.5.2
django-environ==0.4.0
django-filter==0.11.0
django-modeltranslation==0.10.2
django-oauth-toolkit==0.8.1
django-rest-swagger==0.2.9
djangorestframework==3.3.2
easy-thumbnails==2.3
ecdsa==0.13
Fabric==1.10.2
filebrowser-safe==0.4.0
future==0.15.2
gnureadline==6.3.3
grappelli-safe==0.4.1
gunicorn==19.4.5
html5lib==0.9999999
httpie==0.9.2
ipython==4.0.1
ipython-genutils==0.1.0
Markdown==2.6.5
mezzanine-api==0.3.0
MySQL-python==1.2.5
oauthlib==1.0.3
paramiko==1.16.0
path.py==8.1.2
pep8==1.6.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
pycrypto==2.6.1
pyflakes==1.0.0
Pygments==2.0.2
pytz==2015.7
PyYAML==3.11
qiniu==7.0.6
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
tzlocal==1.2
wechat-sdk==0.6.1
xmltodict==0.10.1

